move components to parent directory
