+++ packages/slate-dev-benchmark/src/Timer.js
