--- packages/slate-dev-benchmark/src/components/types.js
@@ -1,10 +1,0 @@
-const RepositoryType = '@@__slate_Repository__@@'
-const SuiteType = '@@__slate_Suite__@@'
-const BenchType = '@@__slate_Bench__@@'
-const TimerType = '@@__slate_Bench_timer_@@'
-module.exports = {
-  RepositoryType,
-  SuiteType,
-  BenchType,
-  TimerType,
-}
