+++ packages/slate-dev-benchmark/src/index.js
@@ -1,5 +1,5 @@
-const { Repository, repo } = require('./components/Repository')
-const { Suite } = require('./components/Suite')
-const { Bench } = require('./components/Bench')
+const { Repository, repo } = require('./Repository')
+const { Suite } = require('./Suite')
+const { Bench } = require('./Bench')
 
 module.exports = { Repository, Suite, Bench, repo }
