+++ packages/slate-dev-benchmark/src/makeOptions.js
