+++ packages/slate-dev-benchmark/src/types.js
@@ -1,0 +1,10 @@
+const RepositoryType = '@@__Slate_Repository__@@'
+const SuiteType = '@@__Slate_Suite__@@'
+const BenchType = '@@__Slate_Bench__@@'
+const TimerType = '@@__Slate_Bench_Timer_@@'
+module.exports = {
+  RepositoryType,
+  SuiteType,
+  BenchType,
+  TimerType,
+}
